chore: clear the clippy backlog across components

+++ components/ical/src/lib.rs
@@ -31,7 +31,7 @@ fn civil_from_days(days: i64) -> (i64, u32, u32) {
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
     let mp = (5 * doy + 2) / 153; // [0, 11]
     let d = (doy - (153 * mp + 2) / 5 + 1) as u32; // [1, 31]
-    let m = if mp < 10 { mp + 3 } else { mp - 9 } as i64; // [1, 12]
+    let m = if mp < 10 { mp + 3 } else { mp - 9 }; // [1, 12]
     (y + if m <= 2 { 1 } else { 0 }, m as u32, d)
 }
